requirements.txt
groq
python-dotenv
sentence-transformers
faiss-cpu